--- simtong-application/src/main/kotlin/team/comit/simtong/domain/user/dto/ChangeEmailRequest.kt
@@ -1,13 +1,0 @@
-package team.comit.simtong.domain.user.dto
-
-/**
- *
- * 이메일 변경 정보를 전달하는 ChangeEmailRequest
- *
- * @author Chokyunghyeon
- * @date 2022/10/03
- * @version 1.0.0
- **/
-data class ChangeEmailRequest(
-    val email: String
-)
