--- simtong-application/src/main/kotlin/team/comit/simtong/domain/user/dto/ChangePasswordRequest.kt
@@ -1,14 +1,0 @@
-package team.comit.simtong.domain.user.dto
-
-/**
- *
- * 사용자의 비밀번호 변경 요청을 하는 ChangePasswordRequest
- *
- * @author Chokyunghyeon
- * @date 2022/10/14
- * @version 1.0.0
- **/
-data class ChangePasswordRequest(
-    val password: String,
-    val newPassword: String
-)
