--- simtong-application/src/main/kotlin/team/comit/simtong/domain/user/dto/CheckMatchedAccountRequest.kt
@@ -1,14 +1,0 @@
-package team.comit.simtong.domain.user.dto
-
-/**
- *
- * 해당 사원번호와 이메일을 가진 계정 여부 확인을 요청하는 CheckMatchedAccountRequest
- *
- * @author Chokyunghyeon
- * @date 2022/10/15
- * @version 1.0.0
- **/
-data class CheckMatchedAccountRequest(
-    val employeeNumber: Int,
-    val email: String
-)
