+++ simtong-application/src/main/kotlin/team/comit/simtong/domain/user/dto/request/ChangeEmailData.kt
@@ -1,0 +1,13 @@
+package team.comit.simtong.domain.user.dto.request
+
+/**
+ *
+ * 이메일 변경 정보를 전달하는 ChangeEmailData
+ *
+ * @author Chokyunghyeon
+ * @date 2022/10/03
+ * @version 1.2.5
+ **/
+data class ChangeEmailData(
+    val email: String
+)
