+++ simtong-application/src/main/kotlin/team/comit/simtong/domain/user/dto/request/ChangeProfileImageData.kt
@@ -1,0 +1,13 @@
+package team.comit.simtong.domain.user.dto.request
+
+/**
+ *
+ * 프로필 사진 변경 요청 정보를 전달하는 ChangeProfileImageData
+ *
+ * @author Chokyunghyeon
+ * @date 2022/10/03
+ * @version 1.2.5
+ **/
+data class ChangeProfileImageData(
+    val profileImagePath: String
+)
